+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove no longer needed "amdgpu.sg_display=0" workaround.
 - Change 'linux-t2' to be the default Linux kernel instead of 'linux-lts'.
+- Change EFI variables to be available again on non-Mac hardware.
 
 ## 3.4.0 - The Game Mode Update